requirements.txt
azure-ai-formrecognizer==3.3.2
azure-core==1.30.0
fastapi==0.96.1
firebase-admin==6.2.0
google-ai-generativelanguage==0.6.5
google-generativeai==0.7.0
groq==0.9.0
openai==1.43.0
prefect==2.20.3
Pillow==10.3.0
PyMuPDF==1.23.26
python-dotenv==1.0.0
uvicorn==0.30.6
